Split compiler includes into .cpp and .hpp pairs
